requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pydantic==2.8.2
openai==1.40.0
anthropic==0.34.1
python-dotenv==1.0.1
